requirements.txt
Flask
numpy
requests
pytesseract
pillow
pymupdf
python-docx
scikit-learn
opencv-python
python-pptx
